Finalize Transposition table and move it in its own file.
